Refactor to move validators to own file
